Rating: half-point increments (0.5)
Rating was integer-only — now supports 0.5 / 1.5 / 2.5 etc. Each dot
splits into two clickable halves; tap the left half for X.5, the right
half for X+1. Tap-same-to-clear still works.

src/lib/components/Rating.svelte
- Each dot becomes a circular container with two flex-children halves.
  Parent has rounded-full + overflow-hidden so the rectangular halves
  clip to a half-circle on each side.
- leftFilled(i) = value >= i + 0.5; rightFilled(i) = value >= i + 1.
  Display mode renders spans; interactive mode renders buttons with
  per-half click handlers.

src/lib/db/types.ts
- Removed .int() from the rating schema. Range stays 0..5.

supabase/migrations/0003_rating_decimals.sql
- ALTER COLUMN rating TYPE numeric USING rating::numeric. Existing
  integer ratings cast losslessly.
- After this migration runs, PostgREST returns rating as a string
  (numeric type behavior); sync-normalize.ts already coerces strings
  in SESSION_NUMERIC_KEYS back to numbers, so no other code changes.

To activate end-to-end:
- Run supabase/migrations/0003_rating_decimals.sql in the Supabase
  SQL editor.
- Hard-refresh the app so the new Rating component renders.

Verified — 564 files type-check (0 errors / 0 warnings), 92/92 tests
pass.

+++ src/lib/db/types.ts
@@ -113,7 +113,8 @@ const sessionBaseShape = {
 	// Milk is required if style === 'latte'; enforcement is at the form layer
 	// (the storage shape stays optional so we can save drafts mid-edit).
 	milk: MilkSchema.optional(),
-	rating: z.number().int().min(0).max(5).optional(),
+	// Half-point scale: 0, 0.5, 1, 1.5, … 5. Form layer enforces the .5 step.
+	rating: z.number().min(0).max(5).optional(),
 	notes: z.string().optional(),
 	createdAt: z.string(),
 	updatedAt: z.string(),
